+++ neodos-kernel/src/main.rs
@@ -252,6 +252,9 @@ pub unsafe extern "sysv64" fn rust_start(boot_info: &BootInfo) -> ! {
         if let Ok(net_id) = object::ob_create_object(ObType::Key, "Network", 10, 0, None) {
             let _ = object::namespace::ob_insert_object("\\Global\\Info\\Network", net_id);
         }
+        if let Ok(proc_id) = object::ob_create_object(ObType::Key, "Process", 12, 0, None) {
+            let _ = object::namespace::ob_insert_object("\\Global\\Info\\Process", proc_id);
+        }
     }
 
     // ============================================